refactor ValuePieEntry move from chart to reports
